Implement comprehensive migration script for tax model updates and inventory adjustments
- Established MongoDB connection and defined migration function.
- Added tax fields to StockItem and removed unused GST fields.
- Converted tax_rate from string to float in StockItem.
- Updated Inventory with tax_rate, hsn_code, and tax_amount from VoucherGST.
- Calculated and added total_tax and total_amount to Voucher.
- Removed unused fields from Voucher and added additional_charge, discount, total, and grand_total.
- Updated grand_total in Voucher from Accounting data.
- Cleaned up unused fields in Inventory and updated total_amount, tax_rate, etc. from StockItem.
- Updated Company, CompanySettings, and Ledger with new tax-related fields.
- Inserted predefined Tax Models if not already present.
- Dropped obsolete GSTRate and VoucherGST collections.

+++ app/database/models/Inventory.py
@@ -61,7 +61,7 @@ class InventoryItemUpdate(BaseModel):
     hsn_code: Optional[str] = None
     unit: Optional[str] = None
     
-    quantity: int
+    quantity: float
     rate: float
     amount: float
     discount_amount: Optional[float] = 0.0
@@ -80,7 +80,7 @@ class CreateInventoryItemWithTAX(BaseModel):
     hsn_code: Optional[str] = None
     unit: Optional[str] = None
     
-    quantity: int
+    quantity: float
     rate: float
     amount: float
     discount_amount: Optional[float] = 0.0
@@ -101,7 +101,7 @@ class UpdateInventoryItemWithTAX(BaseModel):
     hsn_code: Optional[str] = None
     unit: Optional[str] = None
     
-    quantity: int
+    quantity: float
     rate: float
     amount: float
     discount_amount: Optional[float] = 0.0
